transpvisback-master/requirements.txt
asgiref==3.2.10
Django==3.0.7
djangorestframework==3.11.0
djangorestframework-camel-case==1.2.0
django-cors-headers==3.4.0
django-filter==2.3.0
pytz==2020.1
sqlparse==0.3.1
gunicorn==20.0.4
# psycopg2==2.9.5
django-dbbackup==3.3.0
whitenoise==5.0.1
Markdown
requests==2.22.0
scipy==1.9.0
django-heroku==0.3.1
dj-database-url==0.5.0
nltk
numpy==1.20.0
pandas
scikit-learn==1.0.2
tensorflow==2.9.0
spacy-transformers
PyPDF2
fitz
PyMuPDF

